minor bugs fixed on updates views and templates
