+++ leagueoffreljord/userviews.py
@@ -1,7 +1,7 @@
 from django.views.generic import CreateView, View, UpdateView
 from django.contrib.auth.models import User
 from django.urls import reverse_lazy
-from django.shortcuts import render
+from django.shortcuts import render, redirect
 from .usermodels import *
 from .userforms import *
 from leagueoffreljord.userforms import *
@@ -11,58 +11,7 @@ import leagueoffreljord.riotConstants as Consts
 from django.utils.decorators import method_decorator
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
-
-def profile_save(request, profile_form, profile=None):
-    cleaned_data = profile_form.clean()
-    profile = Profile(
-        user=request.user,
-        name=cleaned_data['name'],
-        last_name=cleaned_data['last_name'],
-        dni=cleaned_data['dni'],
-        picture=cleaned_data['picture']
-    )
-    profile.save()
-    return profile
-
-def lolprofile_save(request, lol_profile_form, lol_profile=None):
-    league = 'Unranked'
-    division = ''
-    cleaned_data = lol_profile_form.clean()
-    api = RiotAPI(Consts.KEY)
-    summoner = api.get_summoner_by_name(cleaned_data['nickname'])
-    try:
-        league_data = api.get_league_by_id(summoner['id'])
-    except:
-        messages.error(request, 'El invocador no existe', extra_tags='create-lol-profile-error')
-        return None
-    messages.success(request, 'Perfil de invocador actualizado exitosamente')
-    for queue_data in league_data:
-        if queue_data['queueType'] == Consts.QUEUE['solo']:
-            league = queue_data['tier']
-            division = queue_data['rank']
-    lol_profile = LolProfile(
-        user=request.user,
-        nickname=cleaned_data['nickname'],
-        league=league,
-        division=division,
-        active=True
-    )
-    lol_profile.save()
-    return lol_profile
-
-def get_profile_from_user_or_none(user):
-    try:
-        profile = Profile.objects.get(user=user)
-    except ObjectDoesNotExist:
-        profile = None
-    return profile
-
-def get_lolprofile_from_user_or_none(user):
-    try:
-        lol_profile = LolProfile.objects.get(user=user)
-    except ObjectDoesNotExist:
-        lol_profile = None
-    return lol_profile
+from .useraux import *
 
 class RegistroUsuario(CreateView):
     model = User
@@ -108,7 +57,7 @@ class ProfileView(View):
         if lol_profile_form.is_valid():
             lol_profile = lolprofile_save(request, lol_profile_form)
         else:
-            lol_profile = get_lolprofile_from_user_or_none(user)
+            lol_profile = get_lolprofile_from_user_or_none(request.user)
         context = {
             'profile': profile,
             'lol_profile': lol_profile,
@@ -134,6 +83,7 @@ class ProfileUpdateView(UpdateView):
             messages.success(request, 'Perfil actualizado exitosamente')
         else:
             messages.error(request, 'El formulario es invalido, por favor ingrese los datos nuevamente', extra_tags='update-profile-error')
+        return redirect('api:profile')
 
 class LolProfileUpdateView(UpdateView):
     model = LolProfile
@@ -146,10 +96,13 @@ class LolProfileUpdateView(UpdateView):
 
     @method_decorator(login_required(login_url='api:home', redirect_field_name='api:lol_profile_edit'))
     def post(self, request):
-        lol_profile_form = profile_form = ProfileForm(request.POST, request.FILES)
-        if profile_form.is_valid():
-            lol_profile = get_profile_from_user_or_none(request.user)
-            lol_profile = profile_save(request, lol_profile_form, lol_profile)
+        lol_profile_form = LolProfileForm(request.POST)
+        if lol_profile_form.is_valid():
+            lol_profile = get_lolprofile_from_user_or_none(request.user)
+            lol_profile.active = False
+            lol_profile.save()
+            lol_profile = lolprofile_save(request, lol_profile_form)
             messages.success(request, 'Perfil de invocador actualizado exitosamente')
         else:
             messages.error(request, 'El formulario es invalido, por favor ingrese los datos nuevamente', extra_tags='update-lol-profile-error')
+        return redirect('api:profile')
